Combine director dialogs
